change test cases
